requirements.txt
openai
langchain
tavily-python
arxiv
python-dotenv
langchain_community